Demo swing bug.
In the test folder, run fr.pasteur.iah.swingbug.ShowSwingBug

A GUI for a future TrackMate plugin will show up.
Just select the panel called 'Maximum likelihood-estimation'
and click on start estimation. It will crash but before that
the text fields will show the problem.

You can also see the problem if you click on one of the textfield
with numbres in the same panel.